+++ integration/init_app/integration_test.go
@@ -8,10 +8,8 @@ import (
 	"net/url"
 	"os"
 	"path"
-	"path/filepath"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/docker/docker/client"
 	. "github.com/onsi/ginkgo"
@@ -27,12 +25,12 @@ import (
 )
 
 type TestMetadata struct {
-	CustomerID     string `yaml:"customer_id"`
-	InstallationID string `yaml:"installation_id"`
-	ReleaseVersion string `yaml:"release_version"`
-	SetChannelName string `yaml:"set_channel_name"`
-	Flavor         string `yaml:"flavor"`
-	DisableOnline  bool   `yaml:"disable_online"`
+	InstallationID string   `yaml:"installation_id"`
+	CustomerID     string   `yaml:"customer_id"`
+	ReleaseVersion string   `yaml:"release_version"`
+	SetChannelName string   `yaml:"set_channel_name"`
+	Flavor         string   `yaml:"flavor"`
+	Args           []string `yaml:"args"`
 
 	// debugging
 	SkipCleanup bool `yaml:"skip_cleanup"`
@@ -65,19 +63,8 @@ var _ = Describe("ship init replicated.app/...", func() {
 		if file.IsDir() {
 			When(fmt.Sprintf("the spec in %q is run", file.Name()), func() {
 				testPath := path.Join(integrationDir, file.Name())
-				testInputPath := path.Join(testPath, "input")
 				var testOutputPath string
 				var testMetadata TestMetadata
-				var installationID string
-				customerEndpoint := os.Getenv("SHIP_INTEGRATION_CUSTOMER_ENDPOINT")
-				vendorEndpoint := os.Getenv("SHIP_INTEGRATION_VENDOR_ENDPOINT")
-				vendorToken := os.Getenv("SHIP_INTEGRATION_VENDOR_TOKEN")
-				if customerEndpoint == "" {
-					customerEndpoint = "https://pg.staging.replicated.com/graphql"
-				}
-				if vendorEndpoint == "" {
-					vendorEndpoint = "https://g.staging.replicated.com/graphql"
-				}
 
 				BeforeEach(func(done chan<- interface{}) {
 					os.Setenv("NO_OS_EXIT", "1")
@@ -89,18 +76,15 @@ var _ = Describe("ship init replicated.app/...", func() {
 					// read the test metadata
 					testMetadata = readMetadata(testPath)
 
-					// if a token is provided, try to ensure the release matches what we have here in the repo
-					if vendorToken == "" {
-						Fail("Please set SHIP_INTEGRATION_VENDOR_TOKEN to run the init_app test suite")
-					}
+					// TODO - instead of getting installation ID, etc from test metadata create a release with the vendor api
+					// TODO customer ID and vendor token will need to be read from environment variables
+					// TODO so will the desired environment - staging vs prod
 
-					channelName := fmt.Sprintf("integration replicated.app %s", filepath.Base(testPath))
-					installationID = createRelease(vendorEndpoint, vendorToken, testInputPath, testMetadata, channelName)
 					close(done)
 				}, 20)
 
 				AfterEach(func() {
-					if !testMetadata.SkipCleanup {
+					if !testMetadata.SkipCleanup && os.Getenv("SHIP_INTEGRATION_SKIP_CLEANUP_ALL") == "" {
 						// remove the temporary directory
 						err := os.RemoveAll(testOutputPath)
 						Expect(err).NotTo(HaveOccurred())
@@ -109,21 +93,14 @@ var _ = Describe("ship init replicated.app/...", func() {
 				}, 20)
 
 				It("Should output files matching those expected when communicating with the graphql api", func() {
-					if testMetadata.DisableOnline {
-						Skip("Online test skipped")
-					}
 
-					isStaging := strings.Contains(customerEndpoint, "staging")
-					upstream := "replicated.app/some-cool-ci-tool"
-					if isStaging {
-						upstream = "staging.replicated.app/some-cool-ci-tool"
-					}
+					upstream := "staging.replicated.app/some-cool-ci-tool"
 
 					// this should probably be url encoded but whatever
 					upstream = fmt.Sprintf(
 						"%s?installation_id=%s&customer_id=%s",
 						upstream,
-						installationID,
+						testMetadata.InstallationID,
 						testMetadata.CustomerID,
 					)
 
@@ -134,16 +111,30 @@ var _ = Describe("ship init replicated.app/...", func() {
 						"init",
 						upstream,
 						"--headless",
-						fmt.Sprintf("--state-file=%s", path.Join(testInputPath, ".ship/state.json")),
 						"--log-level=off",
-					}))
+					}, testMetadata.Args...))
 					err := cmd.Execute()
 					Expect(err).NotTo(HaveOccurred())
 
+					// these strings will be replaced in the "expected" yaml before comparison
+					replacements := map[string]string{
+						"__upstream__":       strings.Replace(upstream, "&", "\\u0026", -1), // this string is encoded within the output
+						"__installationID__": testMetadata.InstallationID,
+						"__customerID__":     testMetadata.CustomerID,
+					}
+
 					// compare the files in the temporary directory with those in the "expected" directory
-					result, err := integration.CompareDir(path.Join(testPath, "expected"), testOutputPath)
+					result, err := integration.CompareDir(path.Join(testPath, "expected"), testOutputPath, replacements, []string{}, []map[string][]string{})
 					Expect(err).NotTo(HaveOccurred())
 					Expect(result).To(BeTrue())
+
+					// run 'ship watch' and expect no error to occur
+					watchCmd := cli.RootCmd()
+					watchBuf := new(bytes.Buffer)
+					watchCmd.SetOutput(watchBuf)
+					watchCmd.SetArgs(append([]string{"watch", "--exit"}, testMetadata.Args...))
+					err = watchCmd.Execute()
+					Expect(err).NotTo(HaveOccurred())
 				}, 60)
 			})
 		}
@@ -156,6 +147,7 @@ func createRelease(
 	testInputPath string,
 	testMetadata TestMetadata,
 	channelName string,
+	customerID string,
 ) string {
 	endpointURL, err := url.Parse(vendorEndpoint)
 	Expect(err).NotTo(HaveOccurred())
@@ -175,10 +167,10 @@ func createRelease(
 		string(releaseContents),
 		channel.ID,
 		testMetadata.ReleaseVersion,
-		fmt.Sprintf("integration tests running on %s", time.Now()),
+		"integration tests",
 	)
 	Expect(err).NotTo(HaveOccurred())
-	installationID, err := vendorClient.EnsureCustomerOnChannel(testMetadata.CustomerID, channel.ID)
+	installationID, err := vendorClient.EnsureCustomerOnChannel(customerID, channel.ID)
 	Expect(err).NotTo(HaveOccurred())
 	return installationID
 }
